+++ app/subcommands/repo/clone.sh
@@ -7,7 +7,7 @@ set -euf
 # shellcheck disable=SC1091
 . ./lib.sh
 
-[ -n "${1:-}" ] || fatality "must provide a repo name paramater"
+[ -n "${1:-}" ] || fatality 'must provide a repo name paramater'
 repo="$1"
 url="$(dab config get "repo/$repo/url")"
 [ -n "$url" ] || fatality "url for repo $repo is unknown"